requirements.txt
streamlit==1.32.0
google-generativeai==0.8.6
python-dotenv==1.0.0
openai==1.3.7
supabase
postgrest
psycopg2-binary